move components in to individual folders
