+++ client/src/components/README.md
@@ -1,0 +1,12 @@
+# Components
+
+Components are the basic building blocks of react and we should build them up block by block.
+
+The file structure should be as follows:
+
+- (Folder) `ComponentName`
+  - `ComponentName.tsx`
+  - `ComponentName.story.tsx`
+  - `ComponentName.test.tsx` (optional)
+    - (Folder) `ChildComponent` (optional)
+      - Follows same structure as above
